Refactor: move notification modesl to data.models.notification
