+++ app/src/main/java/fr/free/nrw/commons/data/models/notification/Notification.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.notification
+package fr.free.nrw.commons.data.models.notification
 
 import org.wikipedia.util.DateUtil
 
@@ -6,18 +6,19 @@ import org.wikipedia.util.DateUtil
  * Created by root on 18.12.2017.
  */
 data class Notification(var notificationType: NotificationType,
-                   var notificationText: String,
-                   var date: String,
-                   var link: String,
-                   var iconUrl: String,
-                   var notificationId: String) {
+                        var notificationText: String,
+                        var date: String,
+                        var link: String,
+                        var iconUrl: String,
+                        var notificationId: String) {
 
     companion object {
         @JvmStatic
         fun from(wikiNotification: org.wikipedia.notifications.Notification): Notification {
             val contents = wikiNotification.contents
             val notificationLink = if (contents == null || contents.links == null || contents.links!!.primary == null) "" else contents.links!!.primary!!.url
-            return Notification(NotificationType.UNKNOWN,
+            return Notification(
+                NotificationType.UNKNOWN,
                     contents?.compactHeader ?: "",
                     DateUtil.getMonthOnlyDateString(wikiNotification.timestamp),
                     notificationLink,
