+++ app/src/main/java/fr/free/nrw/commons/data/models/notification/NotificationType.java
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.notification;
+package fr.free.nrw.commons.data.models.notification;
 
 public enum NotificationType {
     THANK_YOU_EDIT("thank-you-edit"),
