+++ app/src/main/java/fr/free/nrw/commons/notification/NotificationClient.java
@@ -1,5 +1,6 @@
 package fr.free.nrw.commons.notification;
 
+import fr.free.nrw.commons.data.models.notification.Notification;
 import org.wikipedia.csrf.CsrfTokenClient;
 import org.wikipedia.dataclient.Service;
